TDLS: Set EHT/MLO information for TDLS STA into the driver
Add the copied EHT capabilities into the sta_add() call when adding a
TDLS peer.

The mld_link_id value was previously only for AP mode, but it can now be
used for TDLS links as well to indicate the link on which a
single-link-TDLS direct link is negotiated.

Signed-off-by: Jouni Malinen <quic_klokere@quicinc.com>

CRs-Fixed: 3616639
Git-repo: git://w1.fi/srv/git/hostap.git
Git-commit: f429064189c3cdbea8689c142fb6e31dc5403a8d
Change-Id: Ib8160c1d147939604b22cd994d55f590dab11ab3