EAP-SIM/AKA peer: Simplify identity selection for MK derivation
Redesign the identity selection for MK derivation to be done explicitly
based on the last indicated identity (whether it is from
EAP-Response/Identity or method specific AT_IDENTITY) during the current
exchange. This makes the implementation cleaner and avoids cases were
more or less duplicated selection steps ended up being slightly
different. This is not as clean as it could otherwise be due to the
exception needed for the IMSI privacy case where the identity used in MK
derivation is actually not the one exchanged in the EAP messages.

Furthermore, this moves the somewhat confusing EAP method specific
tracking of the lasgt EAP-Response/Identity value from EAP-SIM/AKA into
the main EAP peer implementation.

CRs-Fixed: 3622885
Git-repo: git://w1.fi/srv/git/hostap.git
Git-commit: 881cb4198b552562bc732077170219bb295a64e8
Change-Id: Ia7caf47d89271fcff1bcc67a4e5690d77b469fc7
Signed-off-by: Jouni Malinen <quic_jouni@quicinc.com>